requirements.txt
# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Computer Vision & Deep Learning
opencv-python==4.8.1.78
ultralytics==8.1.0  # YOLOv8
deepface==0.0.87
tf-keras==2.15.0
tensorflow==2.15.0

# Image Processing
pillow==10.1.0
numpy==1.24.3

# Utilities
python-dotenv==1.0.0
requests==2.31.0
